refactor(emoji): split emoji styles and addmood styles
