--- src/components/Mood/common.styles.jsx
@@ -1,3 +1,0 @@
-import { css } from 'styled-components';
-
-export {}
