--- src/components/Mood/emojis.js
@@ -1,11 +1,0 @@
-import { ReactComponent as Bad }  from '../Styles/icons/frown.svg';
-import { ReactComponent as Neutral }  from '../Styles/icons/meh.svg';
-import { ReactComponent as Good }  from '../Styles/icons/smile.svg';
-
-const EMOJIS = {
-    bad: Bad,
-    neutral: Neutral,
-    good: Good,
-}
-
-export default EMOJIS;
