+++ src/components/Navigation/index.js
@@ -33,12 +33,12 @@ const NavigationAuth = () => (
 
 const NavigationNonAuth = () => (
     <ul>
-        {/* <li>
+        <li>
             <Link to={ROUTES.LANDING}>Landing</Link>
         </li>
         <li>
             <Link to={ROUTES.SIGN_IN}>Sign In</Link>
-        </li> */}
+        </li>
     </ul>
 );
